pip-requirements.txt
flake8
mypy
pylint

wheel

humanfriendly
python-magic
sqlalchemy
sqlalchemy-stubs
pandas
pandas-stubs
